// features/hero/hero.types.ts
export interface HeroEvent {
  id: number;
  title: string;
  category: string;
  tagline: string;
  description: string;
  date: string;
  fullDate: string;
  time: string;
  venue: string;
  location: string;
  city: string;
  price: string;
  image: string;
  label: string;
  soundPreviewName: string;
}